requirements.txt
accelerate==0.29.3
addict==2.4.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
diffusers==0.27.2
docker-pycreds==0.4.0
dominate==2.9.1
editdistance==0.8.1
einops==0.8.0
fastjsonschema==2.20.0
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
fsspec==2024.3.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
h5py==3.12.1
hf_transfer==0.1.6
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1

ipywidgets==8.1.3
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mediapy==1.2.0
ml-collections==0.1.1
moviepy==1.0.3
mpmath==1.3.0
natsort==8.4.0
nbformat==5.10.4
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.5.0
open3d==0.18.0
opencv-python==4.9.0.80
openexr-python==2024.1.5.11.15.26
pandas==2.2.2
peft==0.12.0
pillow==10.3.0
plotly==5.22.0
plyfile==1.0.3
proglog==0.1.10
protobuf==4.25.3
pyparsing==3.1.2
pyquaternion==0.9.9
PySocks==1.7.1
pysolar==0.11
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
retrying==1.3.4
rpds-py==0.19.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.1
scipy==1.13.0
sentencepiece==0.2.0
sentry-sdk==2.0.1
setproctitle==1.3.3
skylibs==0.7.4
smmap==5.0.1
soupsieve==2.6
sympy==1.12
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.4.24
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.40.1
triton==2.3.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
Werkzeug==3.0.2
widgetsnbextension==4.0.11
xformers==0.0.26.post1
zipp==3.18.1
